requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
matplotlib>=3.7
PyYAML>=6.0
pytest>=8.0
